Working audio and trigger buttons
Audio now works, although it still sounds terrible.
Because of this, added functionality to the controller:
pressing the right trigger will mute and unmute the audio and pressing
the left trigger will cause the diamond to move to a new location
without being "collected" (could be useful in demo if the diamond is in
an unobtainable location).